requirements.txt
streamlit>=1.25.0
python-dotenv>=1.0.0
nltk>=3.8.1
langchain>=0.1.0
langchain-openai>=0.0.23
playwright>=1.41.1
faiss-cpu>=1.7.4
openai>=0.27.0
pandas>=2.0.3
requests>=2.31.0
duckduckgo-search>=0.3.2
